+++ weaviate/collections/classes/config.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from enum import Enum
 from typing import (
     Any,
     ClassVar,
@@ -31,6 +30,7 @@ from weaviate.collections.classes.config_named_vectors import (
     _NamedVectorsUpdate,
 )
 from weaviate.collections.classes.config_vector_index import (
+    _MultiVectorConfigCreate,
     VectorIndexType as VectorIndexTypeAlias,
     VectorFilterStrategy,
 )
@@ -51,8 +51,9 @@ from weaviate.collections.classes.config_vectorizers import VectorDistances as V
 from weaviate.collections.classes.config_vectorizers import Vectorizers as VectorizersAlias
 from weaviate.collections.classes.config_vectorizers import _Vectorizer, _VectorizerConfigCreate
 from weaviate.exceptions import WeaviateInvalidInputError
+from weaviate.str_enum import BaseEnum
 from weaviate.util import _capitalize_first_letter
-from ...warnings import _Warnings
+from weaviate.warnings import _Warnings
 
 # BC for direct imports
 Vectorizers: TypeAlias = VectorizersAlias
@@ -65,7 +66,7 @@ AWSService: TypeAlias = Literal[
 ]
 
 
-class ConsistencyLevel(str, Enum):
+class ConsistencyLevel(str, BaseEnum):
     """The consistency levels when writing to Weaviate with replication enabled.
 
     Attributes:
@@ -79,7 +80,7 @@ class ConsistencyLevel(str, Enum):
     QUORUM = "QUORUM"
 
 
-class DataType(str, Enum):
+class DataType(str, BaseEnum):
     """The available primitive data types in Weaviate.
 
     Attributes:
@@ -121,7 +122,7 @@ class DataType(str, Enum):
     OBJECT_ARRAY = "object[]"
 
 
-class Tokenization(str, Enum):
+class Tokenization(str, BaseEnum):
     """The available inverted index tokenization methods for text properties in Weaviate.
 
     Attributes:
@@ -153,7 +154,7 @@ class Tokenization(str, Enum):
     KAGOME_KR = "kagome_kr"
 
 
-class GenerativeSearches(str, Enum):
+class GenerativeSearches(str, BaseEnum):
     """The available generative search modules in Weaviate.
 
     These modules generate text from text-based inputs.
@@ -174,6 +175,8 @@ class GenerativeSearches(str, Enum):
             Weaviate module backed by FriendliAI generative models.
         `MISTRAL`
             Weaviate module backed by Mistral generative models.
+        `NVIDIA`
+            Weaviate module backed by NVIDIA generative models.
         `OLLAMA`
             Weaviate module backed by generative models deployed on Ollama infrastructure.
         `OPENAI`
@@ -189,12 +192,13 @@ class GenerativeSearches(str, Enum):
     DATABRICKS = "generative-databricks"
     FRIENDLIAI = "generative-friendliai"
     MISTRAL = "generative-mistral"
+    NVIDIA = "generative-nvidia"
     OLLAMA = "generative-ollama"
     OPENAI = "generative-openai"
     PALM = "generative-palm"  # rename to google once all versions support it
 
 
-class Rerankers(str, Enum):
+class Rerankers(str, BaseEnum):
     """The available reranker modules in Weaviate.
 
     These modules rerank the results of a search query.
@@ -211,6 +215,8 @@ class Rerankers(str, Enum):
             Weaviate module backed by VoyageAI reranking models.
         `JINAAI`
             Weaviate module backed by JinaAI reranking models.
+        `NVIDIA`
+            Weaviate module backed by NVIDIA reranking models.
     """
 
     NONE = "none"
@@ -218,9 +224,10 @@ class Rerankers(str, Enum):
     TRANSFORMERS = "reranker-transformers"
     VOYAGEAI = "reranker-voyageai"
     JINAAI = "reranker-jinaai"
+    NVIDIA = "reranker-nvidia"
 
 
-class StopwordsPreset(str, Enum):
+class StopwordsPreset(str, BaseEnum):
     """Preset stopwords to use in the `Stopwords` class.
 
     Attributes:
@@ -234,7 +241,7 @@ class StopwordsPreset(str, Enum):
     EN = "en"
 
 
-class ReplicationDeletionStrategy(str, Enum):
+class ReplicationDeletionStrategy(str, BaseEnum):
     """How object deletions in multi node environments should be resolved.
 
     Attributes:
@@ -249,7 +256,7 @@ class ReplicationDeletionStrategy(str, Enum):
     TIME_BASED_RESOLUTION = "TimeBasedResolution"
 
 
-class PQEncoderType(str, Enum):
+class PQEncoderType(str, BaseEnum):
     """Type of the PQ encoder.
 
     Attributes:
@@ -263,7 +270,7 @@ class PQEncoderType(str, Enum):
     TILE = "tile"
 
 
-class PQEncoderDistribution(str, Enum):
+class PQEncoderDistribution(str, BaseEnum):
     """Distribution of the PQ encoder.
 
     Attributes:
@@ -275,6 +282,17 @@ class PQEncoderDistribution(str, Enum):
 
     LOG_NORMAL = "log-normal"
     NORMAL = "normal"
+
+
+class MultiVectorAggregation(str, BaseEnum):
+    """Aggregation type to use for multivector indices.
+
+    Attributes:
+        `MAX_SIM`
+            Maximum similarity.
+    """
+
+    MAX_SIM = "maxSim"
 
 
 class _PQEncoderConfigCreate(_ConfigCreateModel):
@@ -471,6 +489,16 @@ class _GenerativeMistral(_GenerativeProvider):
     maxTokens: Optional[int]
 
 
+class _GenerativeNvidia(_GenerativeProvider):
+    generative: Union[GenerativeSearches, _EnumLikeStr] = Field(
+        default=GenerativeSearches.NVIDIA, frozen=True, exclude=True
+    )
+    temperature: Optional[float]
+    model: Optional[str]
+    maxTokens: Optional[int]
+    baseURL: Optional[str]
+
+
 class _GenerativeFriendliai(_GenerativeProvider):
     generative: Union[GenerativeSearches, _EnumLikeStr] = Field(
         default=GenerativeSearches.FRIENDLIAI, frozen=True, exclude=True
@@ -625,6 +653,20 @@ class _RerankerVoyageAIConfig(_RerankerProvider):
     model: Optional[Union[RerankerVoyageAIModel, str]] = Field(default=None)
 
 
+class _RerankerNvidiaConfig(_RerankerProvider):
+    reranker: Union[Rerankers, _EnumLikeStr] = Field(
+        default=Rerankers.NVIDIA, frozen=True, exclude=True
+    )
+    model: Optional[str] = Field(default=None)
+    baseURL: Optional[AnyHttpUrl]
+
+    def _to_dict(self) -> Dict[str, Any]:
+        ret_dict = super()._to_dict()
+        if self.baseURL is not None:
+            ret_dict["baseURL"] = self.baseURL.unicode_string()
+        return ret_dict
+
+
 class _Generative:
     """Use this factory class to create the correct object for the `generative_config` argument in the `collections.create()` method.
 
@@ -735,6 +777,31 @@ class _Generative:
                 The maximum number of tokens to generate. Defaults to `None`, which uses the server-defined default
         """
         return _GenerativeMistral(model=model, temperature=temperature, maxTokens=max_tokens)
+
+    @staticmethod
+    def nvidia(
+        *,
+        base_url: Optional[str] = None,
+        model: Optional[str] = None,
+        temperature: Optional[float] = None,
+        max_tokens: Optional[int] = None,
+    ) -> _GenerativeProvider:
+        """
+        Create a `_GenerativeNvidia` object for use when performing AI generation using the `generative-nvidia` module.
+
+        Arguments:
+            `base_url`
+                The base URL where the API request should go. Defaults to `None`, which uses the server-defined default
+            `model`
+                The model to use. Defaults to `None`, which uses the server-defined default
+            `temperature`
+                The temperature to use. Defaults to `None`, which uses the server-defined default
+            `max_tokens`
+                The maximum number of tokens to generate. Defaults to `None`, which uses the server-defined default
+        """
+        return _GenerativeNvidia(
+            model=model, temperature=temperature, maxTokens=max_tokens, baseURL=base_url
+        )
 
     @staticmethod
     def ollama(
@@ -1070,7 +1137,7 @@ class _Reranker:
     ) -> _RerankerProvider:
         """Create a `_RerankerCohereConfig` object for use when reranking using the `reranker-cohere` module.
 
-        See the [documentation](https://weaviate.io/developers/weaviate/modules/retriever-vectorizer-modules/reranker-cohere)
+        See the [documentation](https://weaviate.io/developers/weaviate/model-providers/cohere/reranker)
         for detailed usage.
 
         Arguments:
@@ -1085,7 +1152,7 @@ class _Reranker:
     ) -> _RerankerProvider:
         """Create a `_RerankerJinaAIConfig` object for use when reranking using the `reranker-jinaai` module.
 
-        See the [documentation](https://weaviate.io/developers/weaviate/modules/retriever-vectorizer-modules/reranker-jinaai)
+        See the [documentation](https://weaviate.io/developers/weaviate/model-providers/jinaai/reranker)
         for detailed usage.
 
         Arguments:
@@ -1100,7 +1167,7 @@ class _Reranker:
     ) -> _RerankerProvider:
         """Create a `_RerankerVoyageAIConfig` object for use when reranking using the `reranker-voyageai` module.
 
-        See the [documentation](https://weaviate.io/developers/weaviate/modules/retriever-vectorizer-modules/reranker-voyageai)
+        See the [documentation](https://weaviate.io/developers/weaviate/model-providers/voyageai/reranker)
         for detailed usage.
 
         Arguments:
@@ -1108,6 +1175,24 @@ class _Reranker:
                 The model to use. Defaults to `None`, which uses the server-defined default
         """
         return _RerankerVoyageAIConfig(model=model)
+
+    @staticmethod
+    def nvidia(
+        model: Optional[str] = None,
+        base_url: Optional[AnyHttpUrl] = None,
+    ) -> _RerankerProvider:
+        """Create a `_RerankerNvidiaConfig` object for use when reranking using the `reranker-nvidia` module.
+
+        See the [documentation](https://weaviate.io/developers/weaviate/model-providers/nvidia/reranker)
+        for detailed usage.
+
+        Arguments:
+            `model`
+                The model to use. Defaults to `None`, which uses the server-defined default
+            `baseurl`
+                The base URL to send the reranker requests to. Defaults to `None`, which uses the server-defined default.
+        """
+        return _RerankerNvidiaConfig(model=model, baseURL=base_url)
 
 
 class _CollectionConfigCreateBase(_ConfigCreateModel):
@@ -1508,7 +1593,16 @@ SQConfig = _SQConfig
 
 
 @dataclass
+class _MultiVectorConfig(_ConfigBase):
+    aggregation: str
+
+
+MultiVector = _MultiVectorConfig
+
+
+@dataclass
 class _VectorIndexConfig(_ConfigBase):
+    multi_vector: Optional[_MultiVectorConfig]
     quantizer: Optional[Union[PQConfig, BQConfig, SQConfig]]
 
     def to_dict(self) -> Dict[str, Any]:
@@ -1519,6 +1613,8 @@ class _VectorIndexConfig(_ConfigBase):
             out["bq"] = {**out.pop("quantizer"), "enabled": True}
         elif isinstance(self.quantizer, _SQConfig):
             out["sq"] = {**out.pop("quantizer"), "enabled": True}
+        if self.multi_vector is not None:
+            out["multivector"] = self.multi_vector.to_dict()
         return out
 
 
@@ -1975,6 +2071,16 @@ class _CollectionConfigCreate(_ConfigCreateModel):
         ret_dict["properties"] = existing_props
 
 
+class _VectorIndexMultiVector:
+    @staticmethod
+    def multi_vector(
+        aggregation: Optional[MultiVectorAggregation] = None,
+    ) -> _MultiVectorConfigCreate:
+        return _MultiVectorConfigCreate(
+            aggregation=aggregation.value if aggregation is not None else None,
+        )
+
+
 class _VectorIndexQuantizer:
     @staticmethod
     def pq(
@@ -2039,6 +2145,7 @@ class _VectorIndexQuantizer:
 
 
 class _VectorIndex:
+    MultiVector = _VectorIndexMultiVector
     Quantizer = _VectorIndexQuantizer
 
     @staticmethod
@@ -2050,6 +2157,7 @@ class _VectorIndex:
         return _VectorIndexConfigSkipCreate(
             distance=None,
             quantizer=None,
+            multivector=None,
         )
 
     @staticmethod
@@ -2066,6 +2174,7 @@ class _VectorIndex:
         max_connections: Optional[int] = None,
         vector_cache_max_objects: Optional[int] = None,
         quantizer: Optional[_QuantizerConfigCreate] = None,
+        multi_vector: Optional[_MultiVectorConfigCreate] = None,
     ) -> _VectorIndexConfigHNSWCreate:
         """Create a `_VectorIndexConfigHNSWCreate` object to be used when defining the HNSW vector index configuration of Weaviate.
 
@@ -2087,6 +2196,7 @@ class _VectorIndex:
             maxConnections=max_connections,
             vectorCacheMaxObjects=vector_cache_max_objects,
             quantizer=quantizer,
+            multivector=multi_vector,
         )
 
     @staticmethod
@@ -2106,6 +2216,7 @@ class _VectorIndex:
             distance=distance_metric,
             vectorCacheMaxObjects=vector_cache_max_objects,
             quantizer=quantizer,
+            multivector=None,
         )
 
     @staticmethod
@@ -2123,7 +2234,12 @@ class _VectorIndex:
             See [the docs](https://weaviate.io/developers/weaviate/configuration/indexes#how-to-configure-hnsw) for a more detailed view!
         """  # noqa: D417 (missing argument descriptions in the docstring)
         return _VectorIndexConfigDynamicCreate(
-            distance=distance_metric, threshold=threshold, hnsw=hnsw, flat=flat, quantizer=None
+            distance=distance_metric,
+            threshold=threshold,
+            hnsw=hnsw,
+            flat=flat,
+            quantizer=None,
+            multivector=None,
         )