+++ weaviate/collections/collection/async_.py
@@ -76,7 +76,9 @@ class CollectionAsync(Generic[Properties, References], _CollectionBase[Propertie
 
         self.__cluster = _ClusterAsync(connection)
 
-        self.aggregate = _AggregateCollectionAsync(connection, name, consistency_level, tenant)
+        self.aggregate = _AggregateCollectionAsync(
+            connection, name, consistency_level, tenant, validate_arguments
+        )
         """This namespace includes all the querying methods available to you when using Weaviate's standard aggregation capabilities."""
         self.backup = _CollectionBackupAsync(connection, name)
         """This namespace includes all the backup methods available to you when backing up a collection in Weaviate."""