+++ weaviate/exceptions.py
@@ -3,10 +3,10 @@ Weaviate Exceptions.
 """
 
 from json.decoder import JSONDecodeError
-from typing import Union, Tuple
+from typing import Tuple, Union
 
-from grpc.aio import AioRpcError  # type: ignore
 import httpx
+from grpc.aio import AioRpcError  # type: ignore
 
 ERROR_CODE_EXPLANATION = {
     413: """Payload Too Large. Try to decrease the batch size or increase the maximum request size on your weaviate
@@ -96,7 +96,7 @@ class ResponseCannotBeDecodedError(WeaviateBaseError):
             `response`:
                 The request response of which the status code was unexpected.
         """
-        msg = f"Cannot decode response from weaviate {response} with content {response.text} for request from {location}"
+        msg = f"Cannot decode response from weaviate {response} with content '{response.text}' for request from {location}"
         super().__init__(msg)
         self._status_code: int = response.status_code
 
@@ -371,3 +371,12 @@ class InsufficientPermissionsError(UnexpectedStatusCodeError):
 
     def __init__(self, res: Union[httpx.Response, AioRpcError]) -> None:
         super().__init__("forbidden", res)
+
+
+class WeaviateAgentsNotInstalledError(WeaviateBaseError):
+    """Error raised when trying to use Weaviate Agents without the required dependencies."""
+
+    def __init__(self) -> None:
+        super().__init__(
+            'Weaviate Agents (Alpha) functionality requires additional dependencies. Please install them using: "pip install weaviate-client[agents]"'
+        )