+++ docs/changelog.rst
@@ -1,6 +1,31 @@
 Changelog
 =========
 
+Version 4.11.1
+--------------
+This patch version incldues:
+    - Fixes exporting of named vector collections
+    - Support for upcoming weaviate agents. User ``pip install weaviate-client[agents]`` to try it out
+    - Raise error if a header with value `None`` has been added
+    - Introduce ``collections.use`` as an alias for ``collections.get``
+        - ``collections.use`` is the preferred method to create an in-memory collection object
+        - ``collections.get`` will be deprecated in the future and should not be used in new code
+
+Version 4.11.0
+--------------
+
+This minor version includes:
+    - Support for new 1.29 features:
+        - RBAC is ready for production
+        - Multi-dimensional/COLBert vectors  experimental, breaking changes are expected)
+        - Aggregations are using GRPC now
+    - Improved exception handling for errors returned by Weaviate
+    - Add support for NVIDIA modules:
+        - multi2vec-nvidia
+        - text2vec-nvidia
+        - generative-nvidia
+
+
 Version 4.10.4
 --------------
 This patch version incldues: